♻️이슈 상세 페이지의 사이드바 api 로직 분리 및 ErrorSideBar 컴포넌트 수정 - ErrorSideBar의 props를 issue로 받는게 아닌 contentList를 받도록 수정 - ErrorBoundary 적용을 위해 사이드바의 api 요청을 분리함
